Move plookup time utilities in gemini::plookup
This also disables some warnings about unused variables that are meant
to stay until we implement the psnark verifier. There are other unused
variables related more to the algebraic hashing part, still missing
in both provers.
