+++ src/plookup/mod.rs
@@ -1,1 +1,2 @@
 pub mod streams;
+pub mod time_prover;
